+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/exceptions/UsagePolicyPermissionException.java
@@ -33,15 +33,14 @@ import org.eclipse.tractusx.irs.edc.client.policy.AcceptedPolicy;
  * Usage Policy Permission Exception errors in the contract negotiation.
  */
 @Getter
-public class UsagePolicyPermissionException extends EdcClientException {
+public class UsagePolicyPermissionException extends EdcClientException implements PolicyException {
 
     private final transient Policy policy;
     private final String businessPartnerNumber;
 
     public UsagePolicyPermissionException(final List<AcceptedPolicy> acceptedPolicies,
             final Policy providedCatalogItemPolicy, final String businessPartnerNumber) {
-        super("Policies "
-                + acceptedPolicies.stream().map(policy -> policy.policy().getPolicyId()).toList()
+        super("Policies " + acceptedPolicies.stream().map(policy -> policy.policy().getPolicyId()).toList()
                 + " did not match with policy from " + businessPartnerNumber + ".");
 
         this.policy = providedCatalogItemPolicy;