+++ irs-policy-store/src/main/java/org/eclipse/tractusx/irs/policystore/common/CommonConstants.java
@@ -1,9 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022,2024
- *       2022: ZF Friedrichshafen AG
- *       2022: ISTOS GmbH
- *       2022,2024: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *       2022,2023: BOSCH AG
+ * Copyright (c) 2022,2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
@@ -21,23 +17,15 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.edc.client.exceptions;
+package org.eclipse.tractusx.irs.policystore.common;
 
 /**
- * Exception for errors in the EDC client.
+ * Common constants.
  */
-public class EdcClientException extends Exception {
-
-    public EdcClientException(final String msg, final Throwable cause) {
-        super(msg, cause);
-    }
-
-    public EdcClientException(final Throwable cause) {
-        super(cause);
-    }
-
-    public EdcClientException(final String msg) {
-        super(msg);
-    }
-
+public class CommonConstants {
+    public static final String PROPERTY_BPN = "bpn";
+    public static final String PROPERTY_POLICY_ID = "policyId";
+    public static final String PROPERTY_ACTION = "action";
+    public static final String PROPERTY_CREATED_ON = "createdOn";
+    public static final String PROPERTY_VALID_UNTIL = "validUntil";
 }