+++ irs-policy-store/src/main/java/org/eclipse/tractusx/irs/policystore/models/PolicyWithBpn.java
@@ -1,9 +1,5 @@
 /********************************************************************************
- * Copyright (c) 2022,2024
- *       2022: ZF Friedrichshafen AG
- *       2022: ISTOS GmbH
- *       2022,2024: Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *       2022,2023: BOSCH AG
+ * Copyright (c) 2022,2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  * Copyright (c) 2021,2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
@@ -21,23 +17,16 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.edc.client.exceptions;
+package org.eclipse.tractusx.irs.policystore.models;
+
+import org.eclipse.tractusx.irs.edc.client.policy.Policy;
 
 /**
- * Exception for errors in the EDC client.
+ * Helper record for remembering the BPN the policy belongs to.
+ *
+ * @param bpn    the business partner number
+ * @param policy the policy
  */
-public class EdcClientException extends Exception {
-
-    public EdcClientException(final String msg, final Throwable cause) {
-        super(msg, cause);
-    }
-
-    public EdcClientException(final Throwable cause) {
-        super(cause);
-    }
-
-    public EdcClientException(final String msg) {
-        super(msg);
-    }
+public record PolicyWithBpn(String bpn, Policy policy) {
 
 }