+++ CHANGELOG.md
@@ -8,6 +8,8 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 
 ## [Unreleased]
 
+## [5.2.0] - 2024-07-03
+
 ### Fixed
 
 - Fixed ESS Investigation job processing not starting #579
@@ -35,6 +37,10 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - Support for SingleLevelUsageAsPlanned. #470
 - Documentation to describe the delegate process. #470
 - Added file for CC BY 4.0 license for TRG 7 #681
+- Paging endpoint for Policy Store API: `GET /irs/policies/paged`. #639
+    - Supports multi-sort for the properties "bpn", "validUntil", "policyId", "createdOn", "action" with ascending / descending order.
+    - Supports AND-connected multi-filtering by the properties "bpn", "validUntil", "policyId".
+        - Note: filtering by "createdOn", "validUntil" will be implemented in a subsequent story.
 
 ## [5.1.4] - 2024-05-27
 
@@ -698,7 +704,8 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - **Select Aspects you need**  You are able to select the needed aspects for which you want to collect the correct endpoint information.
 
 
-[Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.4...HEAD
+[Unreleased]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.2.0...HEAD
+[5.2.0]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.4...5.2.0
 [5.1.4]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.3...5.1.4
 [5.1.3]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.2...5.1.3
 [5.1.2]: https://github.com/eclipse-tractusx/item-relationship-service/compare/5.1.1...5.1.2